fix bug

+++ HttpParameters/AddTestStudentVM.cs
@@ -5,7 +5,7 @@
         public int StudentId { get; set; }
         public int SerialNumber { get; set; }
         public int[] TestSubjectId { get; set; }
-        public int[] TheBase { get; set; }//القاعدي
+        public double[] TheBase { get; set; }//القاعدي
         public int[] Additional { get; set; }//الاضافي
         public double[] Mark { get; set; }
     }
